Include the network token balance in the address balance

+++ src/main/java/ch/grignola/service/scanner/common/EthereumTokenBalanceResult.java
@@ -3,6 +3,8 @@ package ch.grignola.service.scanner.common;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigInteger;
+
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class EthereumTokenBalanceResult {
 
@@ -11,5 +13,5 @@ public class EthereumTokenBalanceResult {
     @JsonProperty("message")
     public String message;
     @JsonProperty("result")
-    public long result;
+    public BigInteger result;
 }
